小数点対応のコード追加

+++ database/migrations/2025_01_28_133121_create_market_price_mota_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->unsignedBigInteger('model_name_id');
             $table->unsignedBigInteger('grade_name_id');
             $table->year('year')->nullable(); // 車の年式
-            $table->unsignedInteger('mileage')->nullable(); // 走行距離
+            $table->decimal('mileage', 5, 1)->nullable(); // 走行距離（小数点対応）
             $table->unsignedInteger('min_price')->nullable(); // 最低価格
             $table->unsignedInteger('max_price')->nullable(); // 最高価格
             $table->string('sc_url')->unique(); // スクレイピングURL
